+++ packages/antd/src/AntDValueSelector.tsx
@@ -25,6 +25,7 @@ export const AntDValueSelector = ({
   // Props that should not be in extraProps
   testID: _testID,
   rule: _rule,
+  ruleGroup: _ruleGroup,
   rules: _rules,
   level: _level,
   path: _path,